requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
databases
alembic
pydantic
python-dotenv
psycopg2-binary
httpx
twilio
sendgrid
